Move controller tests into separate namespace
To match the namespace of the controller classes under test.
