+++ tests/Controllers/BoreholeFileControllerTest.cs
@@ -1,5 +1,4 @@
 ﻿using Amazon.S3;
-using BDMS.Controllers;
 using BDMS.Models;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
@@ -12,7 +11,7 @@ using System.Security.Claims;
 using System.Text;
 using static BDMS.Helpers;
 
-namespace BDMS;
+namespace BDMS.Controllers;
 
 [TestClass]
 public class BoreholeFileControllerTest
